Started Implementation on Routing Implementation.
findRoute function takes a NIC and a target MAC address and queries all
connected NIC's recursively to try and find the device with the
specified MAC address. Returns a list of the MAC addresses required to
reach target NIC.

Added JUnit 4 to the project, and started work on a test class for
Routing.

Added toString, equals and hashCode method to MAC class.

Altered NIC getMacAddress to return a MAC object instead of a String.